add:import image param from civitai image with url

+++ app/src/main/java/com/allentom/diffusion/api/civitai/entities/Image.kt
@@ -103,7 +103,7 @@ data class ImageResource(
     val hash: String?,
     val name: String,
     val type: String,
-    val weight: Double?,
+    val weight: Float?,
 )
 
 data class CivitaiResource(
